requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
langchain==0.1.4
langchain-openai==0.0.5
openai==1.10.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
tiktoken==0.5.2
numpy==1.26.3
stripe==7.12.0
# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
google-auth==2.29.0
